+++ src/api/fundingRequestFSForm/index.ts
@@ -1,6 +1,6 @@
 import {FundingRequestFSFormDTO } from "../models";
 import { TableApiBase } from "../tableApiBase";
-const TABLENAME = "x_g_dis_atat_funding_request_fs_form";
+export const TABLENAME = "x_g_dis_atat_funding_request_fs_form";
 export class FundingRequestFSFormApi extends TableApiBase<FundingRequestFSFormDTO> {
   constructor() {
     super(TABLENAME);
