+++ src/services/attachment/index.ts
@@ -1,17 +1,16 @@
 /* eslint-disable camelcase */
 import { TableApiBase } from "@/api/tableApiBase";
 import { AttachmentApi } from "@/api/attachments";
-import { AttachableDTO, AttachmentDTO, BaseTableDTO, FundingPlanDTO } from "@/api/models";
+import { AttachableDTO, AttachmentDTO, BaseTableDTO, 
+  FundingPlanDTO } from "@/api/models";
 import api from "@/api";
 import {
   FundingPlanApi,
   TABLENAME as FundingPlanTableName,
 } from "@/api/fundingPlan";
-
-const getExtension= (filename: string): string => {
-  return filename.substring(filename.lastIndexOf(".") + 1);
-}
-
+import { TABLENAME as FundingRequestFSFormTableName } from "@/api/fundingRequestFSForm";
+import { FundingRequestFSAttachmentService } from "./fundingRequestFSForm";
+import { AttachmentService, AttachmentServiceBase } from "./base";
 
 export const AttachmentServiceCallbacks =  (()=>{
 
@@ -62,11 +61,6 @@ export const AttachmentServiceCallbacks =  (()=>{
 })() 
 
 interface TableAttachment<TModel extends AttachableDTO> {
-  data: TModel;
-  attachment: AttachmentDTO;
-}
-
-interface SnowTableAttachment<TModel extends BaseTableDTO> {
   data: TModel;
   attachment: AttachmentDTO;
 }
@@ -167,94 +161,15 @@ export class FileAttachmentService extends FileAttachmentServiceBase<
 > {}
 
 
-export class AttachmentServiceBase<
-  TTableApi extends TableApiBase<TModel>,
-  TModel extends BaseTableDTO> {
-  attachmentApi = new AttachmentApi();
-  serviceKey: string;
-  tableName: string;
-  tableApi: TTableApi;
-  constructor(serviceKey: string, tableName: string, tableApi: TTableApi) {
-    this.serviceKey = serviceKey;
-    this.tableName = tableName;
-    this.tableApi = tableApi;
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  protected updateRecord(record: TModel, attachmentSysId: string, fileName: string): TModel {
-    throw new Error('not implemented exception');
-  }
-
-  async upload(
-    file: File,
-    onProgress?: (total: number, current: number) => void
-  ): Promise<SnowTableAttachment<TModel>> {
-   
-    //creates initial record
-    let record = await this.tableApi.create();
-
-    if (!record) {
-      throw new Error("failed to create record to associate attachment with");
-    }
-
-    const fileName = file.name;
-    const fileExtension = getExtension(fileName);
-
-    //build attachment object
-    const attachment: AttachmentDTO = {
-      file_name: fileName,
-      table_name: this.tableName,
-      content_type: "*/*",
-      table_sys_id: record.sys_id || "",
-    };
-
-    //upload the Attachment and get the meta data
-    const updatedAttachment = await this.attachmentApi.upload(
-      attachment,
-      file,
-      onProgress
-    );
-
-    AttachmentServiceCallbacks
-      .invokeUploadCallbacks(this.serviceKey, {...attachment, ...updatedAttachment});
-
-    const attachmentSysId = updatedAttachment?.sys_id || "";
-
-    record = this.updateRecord(record, attachmentSysId,fileName);
-    const updatedRecord = await this.tableApi.update(
-      record.sys_id || "",
-      record
-    );
-    
-
-    //return the attachment and table data
-    return {
-      data: updatedRecord,
-      attachment: updatedAttachment,
-    };
-    
-  }
-
-
-  async remove(attachment:AttachmentDTO): Promise<void>{
-
-    if(!attachment){
-      throw new Error('invalid request, attachment required');
-    }
-    //first delete the attachment
-    await this.attachmentApi.remove(attachment.sys_id || "");
-    //then delete the record
-    await this.tableApi.remove(attachment.table_sys_id);
-
-    AttachmentServiceCallbacks.invokeRemoveCallbacks(this.serviceKey, attachment);
-  }
-
-
+export interface RecordManager<TModel extends BaseTableDTO> {
+   retrieveOrCreate: () => Promise<TModel>;
+   save:(record: string) => Promise<void>;
 }
 
 
 export const AttachmentServiceTypes = {
   FundingPlans: "FundingPlans",
+  FundingRequestFSForm: "FundingRequestFSForm"
 };
 
 export const FileAttachmentServiceFactory = (
@@ -273,3 +188,24 @@ export const FileAttachmentServiceFactory = (
     );
   }
 };
+
+export const AttachmentServiceFactory = (attachmentServiceType: string): 
+AttachmentServiceBase<TableApiBase<BaseTableDTO>,BaseTableDTO> => {
+
+  debugger;
+
+  switch(attachmentServiceType){
+
+  case AttachmentServiceTypes.FundingRequestFSForm:
+    return new FundingRequestFSAttachmentService(
+      attachmentServiceType,
+      FundingRequestFSFormTableName,
+      api.fundingRequestFSFormTable);
+
+  default:
+    throw new Error(
+      `unable to create service instance for api ${attachmentServiceType}`
+    );
+    
+  }
+}
